the cohorts disagreed because of the reference, not the montage
This afternoon MESA said plm_offset_aasm moves the PLM count further from the
reference on 6 of 6, p = 0.031. PSG-IPA with twelve scorers says the opposite.
Two things differed at once -- reference and montage -- so neither measurement
could settle it.

Ran PSG-IPA again in MESA's montage: left leg only, no bilateral merge, against
the left-leg annotations of the same twelve scorers. Only the reference then
differs.

                        median deviation      within scorer spread
  bilateral             45 -> 24              0/5 -> 4/5
  single leg            38 -> 38              0/5 -> 3/5

The flag improves agreement in MESA's montage too, so the cohort difference is
not the montage. What remains is the reference: MESA has one scorer, annotates
the left leg only, and recorded zero limb movements on two fully scored nights
-- 965 and 956 events including arousals and respiratory events -- while twelve
PSG-IPA scorers find between 68 and 723 per night. The MESA measurement was
most likely measuring their under-scoring. It does not stand as a refutation.

What does stand: without the flag we find half the PLMs a human counts
(ratio 0.51) and land outside the twelve-scorer spread on all five recordings,
with zero against 13 and 221 on SN3 and SN5. LM detection is reasonable at
0.86 -- the loss is in series detection, and since the human movements go
through the identical series rule, that points at the timing of our movements
rather than the rule.

The flag stays off. n=5, Wilcoxon p=0.19, SN4 overshoots to 687 against a
scorer median of 548, and the PLM index moves enough that flipping it is the
user's call rather than a measurement's.

+++ tests/test_position_min_duration.py
@@ -25,7 +25,7 @@ def _run(minuten_per_code: dict, events_per_code: dict):
     """Bouw een nacht met een opgegeven aantal minuten per positiecode."""
     hypno, pos_epochs = [], []
     for code, minuten in minuten_per_code.items():
-        n_ep = int(round(minuten * 60 / EPOCH_S))
+        n_ep = round(minuten * 60 / EPOCH_S)
         hypno += ["N2"] * n_ep
         pos_epochs += [code] * n_ep
     pos_data = np.repeat(np.array(pos_epochs, dtype=float),
